backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# ── Database ──────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg>=0.31.0
alembic==1.14.0
# psycopg2-binary removed — using asyncpg for all DB access

# ── Validation & Settings ─────────────────────────────────────────────
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic[email]==2.10.4

# ── AI — Pydantic AI ─────────────────────────────────────────────────
pydantic-ai>=0.0.14
openai>=1.50.0
# anthropic>=0.40.0   # uncomment if needed
httpx>=0.27.0

# ── Auth & Security ───────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── Task Queue ────────────────────────────────────────────────────────
celery[redis]==5.4.0
flower==2.0.1
redis==5.2.1

# ── Email ─────────────────────────────────────────────────────────────
aiosmtplib==3.0.2
jinja2==3.1.5          # email HTML templates

# ── Utilities ─────────────────────────────────────────────────────────
python-dateutil==2.9.0
pytz==2024.2
orjson>=3.10.0        # faster JSON serialisation

# ── Dev / Testing ─────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2           # async test client
factory-boy==3.3.1
faker==33.1.0
